simplify layout of storage crate
